Postgres with Docker
In this update, the project have migrated to using Postgres as the DB server from MySQL. Here the Postgres us utilized using a docker image from docker hub that allows the device runs the application does not need to download Postgres but to have installed the docker is enough.
This can be implemented for any kind of dependency that the application rely on, to run them without downloading and installing them individually on the device. And also docker allows to resolve issues in version states.